SIQK/SI: More experimentation for Dave's application.
For Dave's use case, we're trying to get more than just a clipped
polygon. Ideally, it will do a small bit of topology stuff, too, to simplify
post-processing. This is another cut at trying to solve that problem. In this
version, 'inside' is modified to accomplish two things: (1) Polygons are now
open, not closed, sets. (2) A clipping polygon's vertex that is FP identical to
a clip polygon's vertex will definitely be evaluated as outside.

To be clear, none of these details is properly part of clipping. An algorithm
based purely on integration shouldn't care if FP-different but logically
duplicate vertices are in the output vertex list, as long as the output vertex
list describes a polygon that is, up to approximation commensurate with FP,
identical to the correct, infinite-arithmetic polygon. But an algorithm that has
to do a lot of topology stuff will have an easier time if the output vertex list
is free of these types of vertices.