CEDR: Generalize CDR interface for multiple rhos.
In general, there may be n tracer mass fields Qm and m total mass fields
rhom. So far, m has been assumed to be 1. If a problem is feasible that's
fine. But eventually we want to take on infeasible problems. Generalize the CDR
interface for this capability, although we're not yet implementing it. Since
it's not being impl'ed yet, throw if the caller tries to use it.

If-guard some MPI calls.