CAAS: Add optional n_accum_in_place to UserAllReducer.
Reductions sometimes must be done BFB-invariant to rank decomposition;
that is the likeliest motivation for creating a UserAllReducer object. By
default, CAAS provides each DOF to the UserAllReducer. However, DOF order
may allow n DOFs in a row to be accumulated in a BFB-invariant way, e.g.,
if those DOFs are guaranteed always to be on the same processor. If so,
expose that value n here.