Improve mass relative error check.
In the randomized CEDR testing, we permit q < 0. This can lead to arbitrarily
small global mass. Previously, this global mass was used in the denominator of
the relerr global mass check. But that's a bad denominator because it doesn
reflect the actual numerics. Instead, accumulate abs(rho q) and use that as the
denominator. Decrease tol now that we have a better denominator.